chore: Converted preset.js to typescript
